WPCOM: Migrate wpcom.undocumented() imports to wpcom.req
