+++ client/state/imports/actions.js
@@ -1,4 +1,4 @@
-import wpLib from 'calypso/lib/wp';
+import wp from 'calypso/lib/wp';
 import {
 	IMPORTS_AUTHORS_SET_MAPPING,
 	IMPORTS_AUTHORS_START_MAPPING,
@@ -19,8 +19,6 @@ import { appStates } from './constants';
 import { isImporterLocked } from './selectors';
 
 import 'calypso/state/imports/init';
-
-const wpcom = wpLib.undocumented();
 
 const ID_GENERATOR_PREFIX = 'local-generated-id-';
 
@@ -49,6 +47,39 @@ const createImportOrder = ( importerStatus ) =>
 	toApi( {
 		...importerStatus,
 		importerState: appStates.IMPORTING,
+	} );
+
+const updateImporter = ( siteId, importerStatus ) =>
+	wp.req.post( {
+		path: `/sites/${ siteId }/imports/${ importerStatus.importerId }`,
+		formData: [ [ 'importStatus', JSON.stringify( importerStatus ) ] ],
+	} );
+
+const uploadExportFile = ( siteId, params ) =>
+	new Promise( ( resolve, reject ) => {
+		const resolver = ( error, data ) => {
+			error ? reject( error ) : resolve( data );
+		};
+
+		const formData = [
+			[ 'importStatus', JSON.stringify( params.importStatus ) ],
+			[ 'import', params.file ],
+		];
+
+		if ( params.url ) {
+			formData.push( [ 'url', params.url ] );
+		}
+
+		const req = wp.req.post(
+			{
+				path: `/sites/${ siteId }/imports/new`,
+				formData,
+			},
+			resolver
+		);
+
+		req.upload.onprogress = params.onprogress;
+		req.onabort = params.onabort;
 	} );
 
 export const lockImport = ( importerId ) => ( {
@@ -82,12 +113,12 @@ export const cancelImport = ( siteId, importerId ) => async ( dispatch ) => {
 		return;
 	}
 
-	const data = await wpcom.updateImporter( siteId, createCancelOrder( siteId, importerId ) );
+	const data = await updateImporter( siteId, createCancelOrder( siteId, importerId ) );
 	dispatch( receiveImporterStatus( data ) );
 };
 
 export const fetchImporterState = ( siteId ) => async ( dispatch ) => {
-	const data = await wpcom.fetchImporterState( siteId );
+	const data = await wp.req.get( `/sites/${ siteId }/imports/` );
 	dispatch( receiveImporterStatus( data ) );
 };
 
@@ -115,7 +146,7 @@ export const resetImport = ( siteId, importerId ) => async ( dispatch ) => {
 	};
 	dispatch( resetImportAction );
 
-	const data = await wpcom.updateImporter( siteId, createExpiryOrder( siteId, importerId ) );
+	const data = await updateImporter( siteId, createExpiryOrder( siteId, importerId ) );
 	dispatch( receiveImporterStatus( data ) );
 };
 
@@ -130,7 +161,7 @@ export const clearImport = ( siteId, importerId ) => async ( dispatch ) => {
 	};
 	dispatch( resetImportAction );
 
-	const data = await wpcom.updateImporter( siteId, createClearOrder( siteId, importerId ) );
+	const data = await updateImporter( siteId, createClearOrder( siteId, importerId ) );
 	dispatch( receiveImporterStatus( data ) );
 };
 
@@ -173,7 +204,7 @@ export const startImporting = ( importerStatus ) => ( dispatch ) => {
 	};
 	dispatch( startImportingAction );
 
-	return wpcom.updateImporter( siteId, createImportOrder( importerStatus ) );
+	return updateImporter( siteId, createImportOrder( importerStatus ) );
 };
 
 export const setUploadStartState = ( importerId, filenameOrUrl ) => ( {
@@ -190,23 +221,22 @@ export const startUpload = ( importerStatus, file, url = undefined ) => ( dispat
 
 	dispatch( setUploadStartState( importerId, file.name ) );
 
-	return wpcom
-		.uploadExportFile( siteId, {
-			importStatus: toApi( importerStatus ),
-			file,
-			url,
-			onprogress: ( event ) => {
-				dispatch(
-					setUploadProgress( importerId, {
-						uploadLoaded: event.loaded,
-						uploadTotal: event.total,
-					} )
-				);
-			},
-			onabort: () => {
-				dispatch( cancelImport( siteId, importerId ) );
-			},
-		} )
+	return uploadExportFile( siteId, {
+		importStatus: toApi( importerStatus ),
+		file,
+		url,
+		onprogress: ( event ) => {
+			dispatch(
+				setUploadProgress( importerId, {
+					uploadLoaded: event.loaded,
+					uploadTotal: event.total,
+				} )
+			);
+		},
+		onabort: () => {
+			dispatch( cancelImport( siteId, importerId ) );
+		},
+	} )
 		.then( ( data ) => {
 			dispatch( finishUpload( importerId, fromApi( data ) ) );
 		} )
